+++ newclient/src/main.ts
@@ -22,6 +22,13 @@ import i18n from '@app/translations'
 import * as Sentry from '@sentry/browser'
 import { Vue as VueIntegration } from '@sentry/integrations'
 import Vue from 'vue'
+import toChecksum from '@app/core/filters/toChecksum'
+
+router.onError(error => {
+    if (/loading chunk \d* failed./i.test(error.message)) {
+        window.location.reload()
+    }
+})
 
 /*
   ===================================================================================
@@ -146,6 +153,14 @@ Vue.use(Vuetify, {
 
 /*
   ===================================================================================
+    Vue: Filters
+  ===================================================================================
+*/
+
+Vue.filter('toChecksum', toChecksum)
+
+/*
+  ===================================================================================
     Vue: Application Kickstart
   ===================================================================================
 */
@@ -167,11 +182,14 @@ new Vue({
     Sentry
   ===================================================================================
 */
-if (process.env.NODE_ENV === 'production') {
-    const sentryToken = process.env.VUE_APP_SENTRY_SECURITY_TOKEN
+const sentryToken = process.env.VUE_APP_SENTRY_SECURITY_TOKEN
 
-    Sentry.init({
-        dsn: sentryToken,
-        integrations: [new VueIntegration({ Vue, attachProps: true, logErrors: true })]
-    })
-}
+Sentry.init({
+    environment: process.env.NODE_ENV,
+    dsn: sentryToken,
+    integrations: [new VueIntegration({ Vue, attachProps: true, logErrors: true })],
+    beforeSend(event) {
+        return process.env.NODE_ENV === 'production' || process.env.NODE_ENV === 'staging' ? event : null
+    },
+    release: configs.VERSION
+})