+++ newclient/vue.config.js
@@ -3,9 +3,27 @@ const webpack = require('webpack')
 const merge = require('webpack-merge')
 const BundleAnalyzerPlugin = require('webpack-bundle-analyzer').BundleAnalyzerPlugin
 const VuetifyLoaderPlugin = require('vuetify-loader/lib/plugin')
+const version = require('./package.json').version
+const vars = {
+    VERSION: version,
+    ROUTER_MODE: process.env.ROUTER_MODE || 'history'
+}
+
+const sourceMapsConfig = {
+    filename: 'sourcemaps/[file].map'
+}
+sourceMapsConfig.exclude = /vendors.*.*/
 
 const webpackCommon = {
-    plugins: [new VuetifyLoaderPlugin(), new webpack.IgnorePlugin(/^\.\/locale$/, /moment$/)],
+    devtool: false,
+    plugins: [new VuetifyLoaderPlugin(), new webpack.IgnorePlugin(/^\.\/locale$/, /moment$/), new webpack.EnvironmentPlugin(vars)],
+    // module: {
+    //     loaders: [
+    //         {
+    //             loader: 'json'
+    //         }
+    //     ]
+    // },
     resolve: {
         extensions: ['.ts', '.vue', '.json'],
         alias: {
@@ -20,8 +38,7 @@ const webpackCommon = {
 const webpackDevelopment = {}
 
 const webpackProduction = {
-    //  plugins: [new BundleAnalyzerPlugin()],
-    plugins: [],
+    plugins: [new webpack.SourceMapDevToolPlugin(sourceMapsConfig)],
     optimization: {
         namedModules: true,
         moduleIds: 'size',
@@ -53,6 +70,7 @@ const webpackProduction = {
 // }
 
 module.exports = {
+    publicPath: './',
     chainWebpack: config => {
         config.plugin('html').tap(args => {
             args[0].hash = true
@@ -75,7 +93,6 @@ module.exports = {
             .end()
     },
     configureWebpack: process.env.NODE_ENV === 'production' ? merge(webpackCommon, webpackProduction) : webpackCommon,
-    productionSourceMap: false,
     devServer: {
         https: true,
         host: 'localhost',